refactor: move old utils to module
